test: 쿠폰 발급 요청 기능 테스트

+++ src/main/java/com/moabam/api/application/coupon/CouponQueueService.java
@@ -18,7 +18,7 @@ public class CouponQueueService {
 	private final CouponQueueRepository couponQueueRepository;
 
 	public void register(AuthorizationMember member, String couponName) {
-		long registerTime = System.currentTimeMillis();
+		double registerTime = System.currentTimeMillis();
 
 		if (canRegister(couponName)) {
 			log.info("{} 쿠폰이 모두 발급되었습니다.", couponName);
